@@ -5,7 +5,7 @@
   "main": "main",
   "scripts": {
     "test": "jest --verbose",
-    "build": "webpack --mode development",
+    "build": "webpack",
     "watch": "npm run build -- --watch",
     "start": "electron ."
   },
@@ -13,25 +13,22 @@
   "author": "Joe Boyle",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.2.3",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.2.3",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "html-webpack-plugin": "^3.2.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-react-simple": "^1.0.2",
+    "enzyme": "^3.9.0",
     "jest": "^23.6.0",
-    "preact-compat": "^3.18.4",
+    "jest-simple-config": "^0.3.0",
     "prettier": "^1.15.3",
-    "webpack": "^4.28.4",
-    "webpack-cli": "^3.2.1"
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.0",
+    "webpack-simple": "^1.3.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.3.1",
     "axios": "^0.18.0",
     "mbta-client": "^0.4.2",
-    "preact": "^8.4.2",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "styled-components": "^3.4.10"
   }
 }
