@@ -1,4 +1,5 @@
-import { h, render } from 'preact';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import App from './app';
 
-render(<App />, document.body);
+ReactDOM.render(<App />, document.querySelector('.main'));
